+++ src/main/java/com/project/dmsport/domain/notice/presentation/dto/response/NoticeResponse.java
@@ -1,17 +1,17 @@
 package com.project.dmsport.domain.notice.presentation.dto.response;
 
 import com.project.dmsport.domain.notice.domain.enums.NoticeType;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
 import java.time.LocalDateTime;
-import java.util.List;
-
 
 @Getter
-@AllArgsConstructor
-public class QueryAllNoticesResponse {
-
-    private final List<NoticeResponse> notices;
+@Builder
+public class NoticeResponse {
+    private final Long id;
+    private final String title;
+    private final String contentPreview;
+    private final NoticeType type;
+    private final LocalDateTime createdAt;
 }
